replace original PlotsTab with new one
